WIP - refactoring networks module, including sensitivity analysis logic in GrFN and FIB classes
