+++ delphi/GrFN/networks.py
@@ -7,13 +7,17 @@ import inspect
 import json
 import os
 import ast
+import itertools
 
 import torch
+from SALib.analyze import sobol, fast, rbd_fast
+from SALib.sample import saltelli, fast_sampler, latin
 import networkx as nx
 from networkx.algorithms.simple_paths import all_simple_paths
 
+from delphi.GrFN.analysis import get_max_s2_sensitivity
 import delphi.GrFN.utils as utils
-from delphi.GrFN.utils import NodeType, ScopeNode
+from delphi.GrFN.utils import ScopeNode
 from delphi.utils.misc import choose_font
 from delphi.translators.for2py import (
     preprocessor,
@@ -22,6 +26,8 @@ from delphi.translators.for2py import (
     pyTranslate,
     genPGM,
 )
+import numpy as np
+import torch
 
 FONT = choose_font()
 
@@ -29,7 +35,103 @@ dodgerblue3 = "#1874CD"
 forestgreen = "#228b22"
 
 
-class GroundedFunctionNetwork(nx.DiGraph):
+class ComputationalGraph(nx.DiGraph):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.call_graph = self.build_call_graph()
+        self.function_sets = self.build_function_sets()
+        self.outputs = [
+            n
+            for n, d in self.out_degree()
+            if d == 0 and self.nodes[n]["type"] == "variable"
+        ]
+        self.output_node = self.outputs[-1]
+
+    def build_call_graph(self):
+        G = nx.DiGraph()
+        for (name, attrs) in self.nodes(data=True):
+            if attrs["type"] == "function":
+                for predecessor_variable in self.predecessors(name):
+                    for predecessor_function in self.predecessors(
+                        predecessor_variable
+                    ):
+                        G.add_edge(predecessor_function, name)
+
+        return G
+
+    def build_function_sets(self):
+        # TODO - this fails when there is only one function node in the graph -
+        # need to fix!
+        initial_funcs = [n for n, d in self.call_graph.in_degree() if d == 0]
+        distances = dict()
+
+        def find_distances(funcs, dist):
+            all_successors = list()
+            for func in funcs:
+                distances[func] = dist
+                all_successors.extend(self.call_graph.successors(func))
+            if len(all_successors) > 0:
+                find_distances(list(set(all_successors)), dist + 1)
+
+        find_distances(initial_funcs, 0)
+        call_sets = dict()
+        for func_name, call_dist in distances.items():
+            if call_dist in call_sets:
+                call_sets[call_dist].add(func_name)
+            else:
+                call_sets[call_dist] = {func_name}
+
+        function_set_dists = sorted(
+            call_sets.items(), key=lambda t: (t[0], len(t[1]))
+        )
+        function_sets = [func_set for _, func_set in function_set_dists]
+        return function_sets
+
+    def run(
+        self,
+        inputs: Dict[str, Union[float, Iterable]],
+        torch_size: Optional[int] = None,
+    ) -> Union[float, Iterable]:
+        """Executes the GrFN over a particular set of inputs and returns the
+        result.
+
+        Args:
+            inputs: Input set where keys are the names of input nodes in the
+              GrFN and each key points to a set of input values (or just one).
+
+        Returns:
+            A set of outputs from executing the GrFN, one for every set of
+            inputs.
+        """
+        # Abort run if inputs does not match our expected input set
+        if len(inputs) != len(self.inputs):
+            raise ValueError("Incorrect number of inputs.")
+
+        # Set input values
+        for node_name, val in inputs.items():
+            self.nodes[node_name]["value"] = val
+
+        for func_set in self.function_sets:
+            for func_name in func_set:
+                lambda_fn = self.nodes[func_name]["lambda_fn"]
+                output_node = list(self.successors(func_name))[0]
+
+                signature = self.nodes[func_name]["func_inputs"]
+                inputs = [self.nodes[n]["value"] for n in signature]
+                res = lambda_fn(*inputs)
+
+                if torch_size is not None and len(signature) == 0:
+                    self.nodes[output_node]["value"] = torch.tensor(
+                        [res] * torch_size, dtype=torch.double
+                    )
+                else:
+                    self.nodes[output_node]["value"] = res
+
+        # Return the output
+        return self.nodes[self.output_node]["value"]
+
+
+class GroundedFunctionNetwork(ComputationalGraph):
     """
     Representation of a GrFN model as a DiGraph with a set of input nodes and
     currently a single output. The DiGraph is composed of variable nodes and
@@ -46,11 +148,11 @@ class GroundedFunctionNetwork(nx.DiGraph):
     def __init__(self, G, scope_tree):
         super().__init__(G)
         self.scope_tree = scope_tree
-        self.inputs = [n for n, d in self.in_degree() if d == 0 and self.nodes[n]["type"] == "variable"]
-        self.outputs = [n for n, d in self.out_degree() if d == 0 and self.nodes[n]["type"] == "variable"]
-        self.output_node = self.outputs[-1]
-        self.call_graph = self.build_call_graph()
-        self.function_sets = self.build_function_sets()
+        self.inputs = [
+            n
+            for n, d in self.in_degree()
+            if d == 0 and self.nodes[n]["type"] == "variable"
+        ]
 
     def __repr__(self):
         return self.__str__()
@@ -72,12 +174,16 @@ class GroundedFunctionNetwork(nx.DiGraph):
         tab = "  "
         result = list()
         for n in node_set:
-            repr = n \
-                if self.nodes[n]["type"] == "variable" else \
-                f"{n}{inspect.signature(self.nodes[n]['lambda_fn'])}"
+            repr = (
+                n
+                if self.nodes[n]["type"] == "variable"
+                else f"{n}{inspect.signature(self.nodes[n]['lambda_fn'])}"
+            )
 
             result.append(f"{tab * depth}{repr}")
-            result.extend(self.traverse_nodes(self.successors(n), depth=depth+1))
+            result.extend(
+                self.traverse_nodes(self.successors(n), depth=depth + 1)
+            )
         return result
 
     @classmethod
@@ -118,10 +224,7 @@ class GroundedFunctionNetwork(nx.DiGraph):
         scope_tree = nx.DiGraph()
 
         def add_variable_node(
-            basename: str,
-            parent: str,
-            index: int,
-            is_loop_index: bool = False
+            basename: str, parent: str, index: int, is_loop_index: bool = False
         ):
             G.add_node(
                 f"{parent}::{basename}_{index}",
@@ -146,7 +249,9 @@ class GroundedFunctionNetwork(nx.DiGraph):
             else:
                 return find_correct_scope(new_scope, var_name)
 
-        def process_container(scope, loop_index_variable: Optional[str] = None):
+        def process_container(
+            scope, loop_index_variable: Optional[str] = None
+        ):
             for stmt in scope.body:
                 if "name" in stmt:
                     stmt_type = functions[stmt["name"]]["type"]
@@ -155,25 +260,34 @@ class GroundedFunctionNetwork(nx.DiGraph):
                             if "body" in functions[stmt["name"]]:
                                 stmt_type = "literal"
 
-                        output = stmt['output']
+                        output = stmt["output"]
                         scope.variables[output["variable"]] = output["index"]
-                        add_variable_node(output["variable"], scope.name, output["index"])
-                        G.add_edge(stmt["name"],
-                                f"{scope.name}::{output['variable']}_{output['index']}")
+                        add_variable_node(
+                            output["variable"], scope.name, output["index"]
+                        )
+                        G.add_edge(
+                            stmt["name"],
+                            f"{scope.name}::{output['variable']}_{output['index']}",
+                        )
 
                         ordered_inputs = list()
                         for input in stmt.get("input", []):
                             parent = scope.name
                             index = input["index"]
                             base_name = input["variable"]
-                            if index == -1 and base_name != loop_index_variable:
-                                parent, index = find_correct_scope(scope, base_name)
+                            if (
+                                index == -1
+                                and base_name != loop_index_variable
+                            ):
+                                parent, index = find_correct_scope(
+                                    scope, base_name
+                                )
 
                             add_variable_node(
                                 base_name,
                                 parent,
                                 index,
-                                base_name == loop_index_variable
+                                base_name == loop_index_variable,
                             )
                             node_name = f"{parent}::{base_name}_{index}"
                             ordered_inputs.append(node_name)
@@ -191,17 +305,25 @@ class GroundedFunctionNetwork(nx.DiGraph):
                         )
 
                     elif stmt_type == "loop_plate":
-                        index_variable = functions[stmt["name"]]["index_variable"]
+                        index_variable = functions[stmt["name"]][
+                            "index_variable"
+                        ]
                         scope_tree.add_node(stmt["name"], color="blue")
                         scope_tree.add_edge(scope.name, stmt["name"])
-                        new_scope = ScopeNode(functions[stmt["name"]], parent=scope)
-                        process_container(new_scope, loop_index_variable=index_variable)
+                        new_scope = ScopeNode(
+                            functions[stmt["name"]], parent=scope
+                        )
+                        process_container(
+                            new_scope, loop_index_variable=index_variable
+                        )
                     else:
                         pass
                 elif "function" in stmt and stmt["function"] != "print":
                     scope_tree.add_node(stmt["function"], color="green")
                     scope_tree.add_edge(scope.name, stmt["function"])
-                    new_scope = ScopeNode(functions[stmt["function"]], parent=scope)
+                    new_scope = ScopeNode(
+                        functions[stmt["function"]], parent=scope
+                    )
                     process_container(new_scope)
 
         root = data["start"]
@@ -211,24 +333,36 @@ class GroundedFunctionNetwork(nx.DiGraph):
         return cls(G, scope_tree)
 
     @classmethod
-    def from_python_file(cls, python_file, lambdas_path, json_filename: str, stem: str):
+    def from_python_file(
+        cls, python_file, lambdas_path, json_filename: str, stem: str
+    ):
         """Builds GrFN object from Python file."""
         with open(python_file, "r") as f:
             pySrc = f.read()
         return cls.from_python_src(pySrc, lambdas_path, json_filename, stem)
 
     @classmethod
-    def from_python_src(cls, pySrc, lambdas_path, json_filename: str, stem: str, save_file: bool=False):
+    def from_python_src(
+        cls,
+        pySrc,
+        lambdas_path,
+        json_filename: str,
+        stem: str,
+        save_file: bool = False,
+    ):
         """Builds GrFN object from Python source code."""
         asts = [ast.parse(pySrc)]
         pgm_dict = genPGM.create_pgm_dict(
-            lambdas_path, asts, json_filename, {"FileName": f"{stem}.py"},  # HACK
+            lambdas_path,
+            asts,
+            json_filename,
+            {"FileName": f"{stem}.py"},  # HACK
         )
         lambdas = importlib.__import__(stem + "_lambdas")
         return cls.from_dict(pgm_dict, lambdas)
 
     @classmethod
-    def from_fortran_file(cls, fortran_file: str, tmpdir: str="."):
+    def from_fortran_file(cls, fortran_file: str, tmpdir: str = "."):
         """Builds GrFN object from a Fortran program."""
         stem = Path(fortran_file).stem
         if tmpdir == "." and "/" in fortran_file:
@@ -272,127 +406,158 @@ class GroundedFunctionNetwork(nx.DiGraph):
             elif self.nodes[n]["type"] == "function":
                 self.nodes[n]["func_visited"] = False
 
-    def build_call_graph(self):
-        G = nx.DiGraph()
-        for (name, attrs) in self.nodes(data=True):
-            if attrs["type"] == "function":
-                for predecessor_variable in self.predecessors(name):
-                    for predecessor_function in self.predecessors(predecessor_variable):
-                        G.add_edge(predecessor_function, name)
-
-        return G
-
-    def build_function_sets(self):
-        # TODO - this fails when there is only one function node in the graph -
-        # need to fix!
-        initial_funcs = [n for n, d in self.call_graph.in_degree() if d == 0]
-        distances = dict()
-
-        def find_distances(funcs, dist):
-            all_successors = list()
-            for func in funcs:
-                distances[func] = dist
-                all_successors.extend(self.call_graph.successors(func))
-            if len(all_successors) > 0:
-                find_distances(list(set(all_successors)), dist+1)
-
-        find_distances(initial_funcs, 0)
-        call_sets = dict()
-        for func_name, call_dist in distances.items():
-            if call_dist in call_sets:
-                call_sets[call_dist].add(func_name)
+    def sobol_analysis(
+        self, num_samples, prob_def, use_torch=False, var_types=None
+    ):
+        def create_input_tensor(name, samples):
+            type_info = var_types[name]
+            if type_info[0] == str:
+                (val1, val2) = type_info[1]
+                return np.where(samples >= 0.5, val1, val2)
             else:
-                call_sets[call_dist] = {func_name}
+                return torch.tensor(samples)
 
-        function_set_dists = sorted(call_sets.items(), key=lambda t: (t[0], len(t[1])))
-        function_sets = [func_set for _, func_set in function_set_dists]
-        return function_sets
+        samples = saltelli.sample(
+            prob_def, num_samples, calc_second_order=True
+        )
+        if use_torch:
+            samples = np.split(samples, samples.shape[1], axis=1)
+            samples = [s.squeeze() for s in samples]
+            if var_types is None:
+                values = {
+                    n: torch.tensor(s)
+                    for n, s in zip(prob_def["names"], samples)
+                }
+            else:
+                values = {
+                    n: create_input_tensor(n, s)
+                    for n, s in zip(prob_def["names"], samples)
+                }
+            Y = self.run(values, torch_size=len(samples[0])).numpy()
+        else:
+            Y = np.zeros(samples.shape[0])
+            for i, sample in enumerate(samples):
+                values = {n: val for n, val in zip(prob_def["names"], sample)}
+                Y[i] = self.run(values)
 
-    def run(self, inputs: Dict[str, Union[float, Iterable]], torch_size: Optional[int] = None) -> Union[float, Iterable]:
-        """Executes the GrFN over a particular set of inputs and returns the
-        result.
+        return sobol.analyze(prob_def, Y, print_to_console=True)
+
+    def S2_surface(self, sizes, bounds, presets, use_torch=False):
+        """Calculates the sensitivity surface of a GrFN for the two variables with
+        the highest S2 index.
 
         Args:
-            inputs: Input set where keys are the names of input nodes in the
-              GrFN and each key points to a set of input values (or just one).
+            num_samples: Number of samples for sensitivity analysis.
+            sizes: Tuple of (number of x inputs, number of y inputs).
+            bounds: Set of bounds for GrFN inputs.
+            presets: Set of standard values for GrFN inputs.
 
         Returns:
-            A set of outputs from executing the GrFN, one for every set of
-            inputs.
+            Tuple:
+                Tuple: The names of the two variables that were selected
+                Tuple: The X, Y vectors of eval values
+                Z: The numpy matrix of output evaluations
+
         """
-        # Abort run if inputs does not match our expected input set
-        if len(inputs) != len(self.inputs):
-            raise ValueError("Incorrect number of inputs.")
+        args = self.inputs
+        Si = self.sobol_analysis(
+            10,
+            {
+                "num_vars": len(args),
+                "names": args,
+                "bounds": [bounds[arg] for arg in args],
+            },
+        )
+        S2 = Si["S2"]
+        (s2_max, v1, v2) = get_max_s2_sensitivity(S2)
 
-        # Set input values
-        for node_name, val in inputs.items():
-            self.nodes[node_name]["value"] = val
+        x_var = args[v1]
+        y_var = args[v2]
+        search_space = [(x_var, bounds[x_var]), (y_var, bounds[y_var])]
+        preset_vals = {
+            arg: presets[arg]
+            for i, arg in enumerate(args)
+            if i != v1 and i != v2
+        }
 
-        for func_set in self.function_sets:
-            for func_name in func_set:
-                lambda_fn = self.nodes[func_name]["lambda_fn"]
-                output_node = list(self.successors(func_name))[0]
+        X = np.linspace(*search_space[0][1], sizes[0])
+        Y = np.linspace(*search_space[1][1], sizes[1])
 
-                signature = self.nodes[func_name]["func_inputs"]
-                inputs = [self.nodes[n]["value"] for n in signature]
-                res = lambda_fn(*inputs)
+        if use_torch:
+            Xm, Ym = torch.meshgrid(torch.tensor(X), torch.tensor(Y))
+            inputs = {n: torch.full_like(Xm, v) for n, v in presets.items()}
+            inputs.update({search_space[0][0]: Xm, search_space[1][0]: Ym})
+            Z = self.run(inputs).numpy()
+        else:
+            Xm, Ym = np.meshgrid(X, Y)
+            Z = np.zeros((len(X), len(Y)))
+            for x, y in itertools.product(range(len(X)), range(len(Y))):
+                inputs = {n: v for n, v in presets.items()}
+                inputs.update({search_space[0][0]: x, search_space[1][0]: y})
+                Z[x][y] = self.run(inputs)
 
-                if torch_size is not None and len(signature) == 0:
-                    self.nodes[output_node]["value"] = torch.tensor([res] * torch_size, dtype=torch.double)
-                else:
-                    self.nodes[output_node]["value"] = res
-
-        # return the output
-        return self.nodes[self.output_node]["value"]
+        return X, Y, Z, x_var, y_var
 
     def to_CAG(self):
         G = nx.DiGraph()
         for (name, attrs) in self.nodes(data=True):
             if attrs["type"] == "variable":
                 for pred_fn in self.predecessors(name):
-                    if not any(fn_type in pred_fn for fn_type in ("condition", "decision")):
+                    if not any(
+                        fn_type in pred_fn
+                        for fn_type in ("condition", "decision")
+                    ):
                         for pred_var in self.predecessors(pred_fn):
-                            G.add_node(self.nodes[pred_var]['basename'], **self.nodes[pred_var])
-                            G.add_node(attrs['basename'], **attrs)
+                            G.add_node(
+                                self.nodes[pred_var]["basename"],
+                                **self.nodes[pred_var],
+                            )
+                            G.add_node(attrs["basename"], **attrs)
                             G.add_edge(
-                                self.nodes[pred_var]['basename'],
-                                attrs['basename']
+                                self.nodes[pred_var]["basename"],
+                                attrs["basename"],
                             )
                 if attrs["is_loop_index"]:
-                    G.add_edge(attrs['basename'], attrs['basename'])
+                    G.add_edge(attrs["basename"], attrs["basename"])
 
         return G
 
     def to_FIB(self, other):
         if not isinstance(other, GroundedFunctionNetwork):
-            raise TypeError(f"Expected GroundedFunctionNetwork, but got {type(other)}")
+            raise TypeError(
+                f"Expected GroundedFunctionNetwork, but got {type(other)}"
+            )
 
         def shortname(var):
-            return var[var.find("::") + 2: var.rfind("_")]
+            return var[var.find("::") + 2 : var.rfind("_")]
 
         def shortname_vars(graph, shortname):
             return [v for v in graph.nodes() if shortname in v]
 
-        this_var_nodes = [shortname(n) for (n, d) in self.nodes(data=True)
-                          if d["type"] == "variable"]
-        other_var_nodes = [shortname(n) for (n, d) in other.nodes(data=True)
-                           if d["type"] == "variable"]
+        this_var_nodes = [
+            shortname(n)
+            for (n, d) in self.nodes(data=True)
+            if d["type"] == "variable"
+        ]
+        other_var_nodes = [
+            shortname(n)
+            for (n, d) in other.nodes(data=True)
+            if d["type"] == "variable"
+        ]
 
         shared_vars = set(this_var_nodes).intersection(set(other_var_nodes))
-        full_shared_vars = {full_var for shared_var in shared_vars
-                            for full_var in shortname_vars(self, shared_var)}
+        full_shared_vars = {
+            full_var
+            for shared_var in shared_vars
+            for full_var in shortname_vars(self, shared_var)
+        }
 
         return ForwardInfluenceBlanket(self, full_shared_vars)
 
     def to_agraph(self):
         A = nx.nx_agraph.to_agraph(self)
         A.graph_attr.update(
-            {
-                "dpi": 227,
-                "fontsize": 20,
-                "fontname": "Menlo",
-                "rankdir": "TB",
-            }
+            {"dpi": 227, "fontsize": 20, "fontname": "Menlo", "rankdir": "TB"}
         )
         A.node_attr.update({"fontname": "Menlo"})
 
@@ -425,71 +590,63 @@ class GroundedFunctionNetwork(nx.DiGraph):
         """
         CAG = self.to_CAG()
         A = nx.nx_agraph.to_agraph(CAG)
-        A.graph_attr.update({
-            "dpi": 227,
-            "fontsize": 20,
-            "fontname": "Menlo"
-        })
-        A.node_attr.update({
-            "shape": "rectangle",
-            "color": "#650021",
-            "style": "rounded",
-            "fontname": "Gill Sans",
-        })
-        A.edge_attr.update({
-            "color": "#650021",
-            "arrowsize": 0.5
-        })
+        A.graph_attr.update({"dpi": 227, "fontsize": 20, "fontname": "Menlo"})
+        A.node_attr.update(
+            {
+                "shape": "rectangle",
+                "color": "#650021",
+                "style": "rounded",
+                "fontname": "Gill Sans",
+            }
+        )
+        A.edge_attr.update({"color": "#650021", "arrowsize": 0.5})
         return A
 
     def to_call_agraph(self):
         A = nx.nx_agraph.to_agraph(self.call_graph)
-        A.graph_attr.update({
-            "dpi": 227,
-            "fontsize": 20,
-            "fontname": "Menlo"
-        })
-        A.node_attr.update({
-            "shape": "rectangle",
-            "color": "#650021",
-            "style": "rounded",
-        })
-        A.edge_attr.update({
-            "color": "#650021",
-            "arrowsize": 0.5
-        })
+        A.graph_attr.update({"dpi": 227, "fontsize": 20, "fontname": "Menlo"})
+        A.node_attr.update(
+            {"shape": "rectangle", "color": "#650021", "style": "rounded"}
+        )
+        A.edge_attr.update({"color": "#650021", "arrowsize": 0.5})
         return A
 
 
-class ForwardInfluenceBlanket(nx.DiGraph):
+class ForwardInfluenceBlanket(ComputationalGraph):
     """
     This class takes a network and a list of a shared nodes between the input
     network and a secondary network. From this list a shared nodes and blanket
     network is created including all of the nodes between any input/output pair
     in the shared nodes, as well as all nodes required to blanket the network
     for forward influence. This class itself becomes the blanket and inherits
-    from the NetworkX DiGraph class.
+    from the ComputationalGraph class.
     """
+
     def __init__(self, G: GroundedFunctionNetwork, shared_nodes: Set[str]):
-        super(ForwardInfluenceBlanket, self).__init__()
+        super().__init__(G)
         self.inputs = set(G.inputs).intersection(shared_nodes)
-        self.output_node = G.output_node
 
         # Get all paths from shared inputs to shared outputs
         path_inputs = shared_nodes - {self.output_node}
         io_pairs = [(inp, self.output_node) for inp in path_inputs]
-        paths = [p for (i, o) in io_pairs for p in all_simple_paths(G, i, o)]
+        paths = [
+            p for (i, o) in io_pairs for p in all_simple_paths(self, i, o)
+        ]
 
         # Get all edges needed to blanket the included nodes
         main_nodes = {node for path in paths for node in path}
-        main_edges = {(n1, n2) for p in paths for n1, n2 in zip(p, p[1:])}
+        main_edges = {
+            (n1, n2) for path in paths for n1, n2 in zip(path, path[1:])
+        }
         self.cover_nodes = set()
         add_nodes, add_edges = list(), list()
 
         def place_var_node(var_node):
-            prev_funcs = list(G.predecessors(var_node))
-            if len(prev_funcs) > 0 and \
-               G.nodes[prev_funcs[0]]["label"] == "L":
+            prev_funcs = list(self.predecessors(var_node))
+            if (
+                len(prev_funcs) > 0
+                and self.nodes[prev_funcs[0]]["label"] == "L"
+            ):
                 prev_func = prev_funcs[0]
                 add_nodes.extend([var_node, prev_func])
                 add_edges.append((prev_func, var_node))
@@ -497,15 +654,15 @@ class ForwardInfluenceBlanket(nx.DiGraph):
                 self.cover_nodes.add(var_node)
 
         for node in main_nodes:
-            if G.nodes[node]["type"] == "function":
-                for var_node in G.predecessors(node):
+            if self.nodes[node]["type"] == "function":
+                for var_node in self.predecessors(node):
                     if var_node not in main_nodes:
                         add_edges.append((var_node, node))
                         if "::IF_" in var_node:
-                            if_func = list(G.predecessors(var_node))[0]
+                            if_func = list(self.predecessors(var_node))[0]
                             add_nodes.extend([if_func, var_node])
                             add_edges.append((if_func, var_node))
-                            for new_var_node in G.predecessors(if_func):
+                            for new_var_node in self.predecessors(if_func):
                                 add_edges.append((new_var_node, if_func))
                                 place_var_node(new_var_node)
                         else:
@@ -514,89 +671,28 @@ class ForwardInfluenceBlanket(nx.DiGraph):
         main_nodes |= set(add_nodes)
         main_edges |= set(add_edges)
         main_nodes = main_nodes - self.inputs - {self.output_node}
-        orig_nodes = G.nodes(data=True)
 
-        self.add_nodes_from([(n, d) for n, d in orig_nodes if n in self.inputs])
         for node in self.inputs:
             self.nodes[node]["color"] = dodgerblue3
             self.nodes[node]["fontcolor"] = dodgerblue3
-            self.nodes[node]["fontname"] = FONT
             self.nodes[node]["penwidth"] = 3.0
 
-        self.add_nodes_from([(n, d) for n, d in orig_nodes
-                            if n in self.cover_nodes])
         for node in self.cover_nodes:
             self.nodes[node]["color"] = forestgreen
             self.nodes[node]["fontcolor"] = forestgreen
-            self.nodes[node]["fontname"] = FONT
 
-        self.add_nodes_from([(n, d) for n, d in orig_nodes if n in main_nodes])
-        for node in main_nodes:
-            self.nodes[node]["fontname"] = FONT
-
-        self.add_node(self.output_node, **G.nodes[self.output_node])
         self.nodes[self.output_node]["color"] = dodgerblue3
         self.nodes[self.output_node]["fontcolor"] = dodgerblue3
-        self.nodes[self.output_node]["fontname"] = FONT
 
         self.add_edges_from(main_edges)
 
-        # NOTE: Adding cut nodes as needed for display only!!
-        # cut_nodes = [n for n in G.nodes if n not in self.nodes]
-        # cut_edges = [e for e in G.edges if e not in self.edges]
-        #
-        # self.add_nodes_from(cut_nodes)
-        # self.add_edges_from(cut_edges)
-        #
-        # for node_name in cut_nodes:
-        #     self.nodes[node_name]["color"] = "orange"
-        #     self.nodes[node_name]["fontcolor"] = "orange"
-        #
-        # for source, dest in cut_edges:
-        #     self[source][dest]["color"] = "orange"
-
-        self.call_graph = self.build_call_graph()
-        self.function_sets = self.build_function_sets()
-
-    def build_call_graph(self):
-        G = nx.DiGraph()
-        for (name, attrs) in self.nodes(data=True):
-            if attrs["type"] == "function":
-                for predecessor_variable in self.predecessors(name):
-                    for predecessor_function in self.predecessors(predecessor_variable):
-                        G.add_edge(predecessor_function, name)
-
-        return G
-
-    def build_function_sets(self):
-        # TODO - this fails when there is only one function node in the graph -
-        # need to fix!
-        initial_funcs = [n for n, d in self.call_graph.in_degree() if d == 0]
-        distances = dict()
-
-        def find_distances(funcs, dist):
-            all_successors = list()
-            for func in funcs:
-                distances[func] = dist
-                all_successors.extend(self.call_graph.successors(func))
-            if len(all_successors) > 0:
-                find_distances(list(set(all_successors)), dist+1)
-
-        find_distances(initial_funcs, 0)
-        call_sets = dict()
-        for func_name, call_dist in distances.items():
-            if call_dist in call_sets:
-                call_sets[call_dist].add(func_name)
-            else:
-                call_sets[call_dist] = {func_name}
-
-        function_set_dists = sorted(call_sets.items(), key=lambda t: (t[0], len(t[1])))
-        function_sets = [func_set for _, func_set in function_set_dists]
-        return function_sets
-
-    @utils.timeit
-    def run(self, inputs: Dict[str, Union[float, Iterable]], covers: Dict[str, Union[float, Iterable]], torch_size: Optional[int] = None) -> Union[float, Iterable]:
-        """Executes the GrFN over a particular set of inputs and returns the
+    def run(
+        self,
+        inputs: Dict[str, Union[float, Iterable]],
+        covers: Dict[str, Union[float, Iterable]],
+        torch_size: Optional[int] = None,
+    ) -> Union[float, Iterable]:
+        """Executes the FIB over a particular set of inputs and returns the
         result.
         Args:
             inputs: Input set where keys are the names of input nodes in the
@@ -613,42 +709,11 @@ class ForwardInfluenceBlanket(nx.DiGraph):
         for node_name, val in covers.items():
             self.nodes[node_name]["value"] = val
 
-        # Abort run if inputs does not match our expected input set
-        if len(inputs) != len(self.inputs):
-            raise ValueError("Incorrect number of inputs.")
-
-        # Set input values
-        for node_name, val in inputs.items():
-            self.nodes[node_name]["value"] = val
-
-        for func_set in self.function_sets:
-            for func_name in func_set:
-                lambda_fn = self.nodes[func_name]["lambda_fn"]
-                output_node = list(self.successors(func_name))[0]
-
-                signature = self.nodes[func_name]["func_inputs"]
-                inputs = [self.nodes[n]["value"] for n in signature]
-                res = lambda_fn(*inputs)
-
-                if torch_size is not None and len(signature) == 0:
-                    self.nodes[output_node]["value"] = torch.tensor([res] * torch_size, dtype=torch.double)
-                else:
-                    self.nodes[output_node]["value"] = res
-
-        # return the output
-        return self.nodes[self.output_node]["value"]
+        return super().run(inputs, torch_size)
 
     def to_agraph(self):
         A = nx.nx_agraph.to_agraph(self)
-        A.graph_attr.update({"dpi": 227, "fontsize": 20, "fontname": "Menlo"})
-        A.node_attr.update({
-            "shape": "rectangle",
-            # "color": "#650021",
-            "style": "rounded",
-            # "fontname": "Gill Sans",
-        })
-        A.edge_attr.update({
-            # "color": "#650021",
-            "arrowsize": 0.5
-        })
+        A.graph_attr.update({"dpi": 227, "fontsize": 20})
+        A.node_attr.update({"shape": "rectangle", "style": "rounded"})
+        A.edge_attr.update({"arrowsize": 0.5})
         return A
